--- src/test/scala-2.11/markets/orderbooks/AbstractOrderBookSpec.scala
@@ -1,37 +1,0 @@
-/*
-Copyright 2016 ScalABM
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-   http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-package markets.orderbooks
-
-import markets.orders.{AskOrder, BidOrder}
-import markets.tradables.Tradable
-import org.scalatest.{FeatureSpec, Matchers}
-
-import scala.util.Random
-
-
-abstract class AbstractOrderBookSpec extends FeatureSpec with Matchers {
-
-  def prng: Random
-
-  val invalidTradable = Tradable("APPL")
-
-  val validTradable = Tradable("GOOG")
-
-  def askOrderBookFactory(tradable: Tradable): AbstractOrderBook[AskOrder]
-
-  def bidOrderBookFactory(tradable: Tradable): AbstractOrderBook[BidOrder]
-
-}
